Pull out name map reading to be more generic
