+++ src/pdbdump/main.cpp
@@ -25,6 +25,7 @@
 #include <string>
 
 #include "msf/msf.h"
+#include "pdb/format.h"
 #include "pdb/pdb.h"
 #include "pdbdump/dump.h"
